requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==10.3.0
celery==5.4.0
redis==5.0.4
boto3==1.34.144
python-dotenv==1.0.1
